Feat: many improvements in the build process

+++ mympdos/build/build.sh
@@ -6,14 +6,30 @@
 #
 
 BUILDDIR="/usr/build"
-POWEROFF="1"
+POWEROFF="0"
+ARCH=$(uname -m)
 
 #Build packages
 B_BUILD="1"
 B_MYMPD="1"
 B_MYMPD_BRANCH="devel"
+B_LIBMPDCLIENT="1"
 B_MPD_STABLE="1"
 B_MPD_MASTER="0"
+
+get_pkgver()
+{
+  PKGVER=$(grep ^pkgver "$1/APKBUILD" | cut -d= -f2)
+  echo "${PKGVER}"
+}
+
+get_pkgname()
+{
+  PKGNAME=$(grep ^pkgname "$1/APKBUILD" | cut -d= -f2)
+  PKGVER=$(grep ^pkgver "$1/APKBUILD" | cut -d= -f2)
+  PKGREL=$(grep ^pkgrel "$1/APKBUILD" | cut -d= -f2)
+  echo "${PKGNAME}-${PKGVER}-r${PKGREL}.apk"
+}
 
 echo ""
 echo "Starting myMPDos build"
@@ -70,51 +86,73 @@ else
 fi
 cp .abuild/*.rsa.pub /etc/apk/keys/
 
-su build -c "install -d packages/package/aarch64/"
+su build -c "install -d packages/package/$ARCH/"
 if [ -f /media/vda1/mympdos-apks/APKINDEX.tar.gz ]
 then
   echo "Restoring existing packages"
-  su build -c "cp /media/vda1/mympdos-apks/* packages/package/aarch64/"
+  su build -c "cp /media/vda1/mympdos-apks/* packages/package/$ARCH/"
 else
   echo "No existing packages found"
 fi
 
+LIBMPDCLIENT_PACKAGE=$(get_pkgname /media/vda1/mympd/libmpdclient)
+B_LIBMPDCLIENT_VER=$(get_pkgver /media/vda1/mympd/libmpdclient)
+if [ "$B_LIBMPDCLIENT" = "1" ] && [ ! -f "packages/package/$ARCH/$LIBMPDCLIENT_PACKAGE" ]
+then
+  echo "Build libmpdclient"
+  su build -c "cp -r /media/vda1/mympd/libmpdclient ."
+  cd libmpdclient || exit 1
+  su build -c "abuild checksum"
+  su build -c "abuild -r"
+  cd ..
+  mv packages/build/"$ARCH"/*.apk "packages/package/$ARCH/"
+fi
+
+echo "/usr/build/packages/package/" >> /etc/apk/repositories
+apk update
+
 if [ "$B_MYMPD" = "1" ]
 then
-  echo "Building myMPD"
+  echo "Build myMPD"
   su build -c "git clone -b $B_MYMPD_BRANCH --depth=1 https://github.com/jcorporation/myMPD.git"
   cd myMPD || exit 1
-  su build -c "./build.sh pkgalpine"
+  MYMPD_PACKAGE=$(get_pkgname contrib/packaging/alpine)
+  if [ ! -f "../packages/package/$ARCH/$MYMPD_PACKAGE" ]
+  then
+    su build -c "./build.sh pkgalpine"
+  fi
   cd ..
 fi
-if [ "$B_MPD_MASTER" = "1" ]
+
+MPD_MASTER_PACKAGE=$(get_pkgname /media/vda1/mympd/mpd-master)
+B_MPD_MASTER_VER=$(get_pkgver /media/vda1/mympd/mpd-master)
+if [ "$B_MPD_MASTER" = "1" ] && [ ! -f "packages/package/$ARCH/$MPD_MASTER_PACKAGE" ]
 then
   echo "Build MDP master"
-  MPDVER="0.22.0"
   su build -c "cp -r /media/vda1/mympd/mpd-master ."
   cd mpd-master || exit 1
-  sed -e "s/__MPDVER__/${MPDVER}/g" -i APKBUILD
   su build -c "git clone -b master --depth=1 https://github.com/MusicPlayerDaemon/MPD.git"
-  mv MPD "mympdos-mpd-master-${MPDVER}"
-  tar -czf mympdos-mpd-master.tar.gz "mympdos-mpd-master-${MPDVER}"
-  rm -fr "mympdos-mpd-master-${MPDVER}"
+  mv MPD "mympdos-mpd-master-${B_MPD_MASTER_VER}"
+  tar -czf mympdos-mpd-master.tar.gz "mympdos-mpd-master-${B_MPD_MASTER_VER}"
+  rm -fr "mympdos-mpd-master-${B_MPD_MASTER_VER}"
   su build -c "abuild checksum"
   su build -c "abuild -r"
   cd ..
 fi
-if [ "$B_MPD_STABLE" = "1" ]
+
+MPD_STABLE_PACKAGE=$(get_pkgname /media/vda1/mympd/mpd-stable)
+B_MPD_STABLE_VER=$(get_pkgver /media/vda1/mympd/mpd-stable)
+if [ "$B_MPD_STABLE" = "1" ] && [ ! -f "packages/package/$ARCH/$MPD_STABLE_PACKAGE" ]
 then
   echo "Building MPD stable"
-  MPDVER="0.21.25"
   su build -c "cp -r /media/vda1/mympd/mpd-stable ."
   cd mpd-stable || exit 1
-  sed -e "s/__MPDVER__/${MPDVER}/g" -i APKBUILD
-  su build -c "wget http://www.musicpd.org/download/mpd/0.21/mpd-${MPDVER}.tar.xz"
-  tar -xzf "mpd-${MPDVER}.tar.xz"
-  rm "mpd-${MPDVER}.tar.xz"
-  mv "mpd-${MPDVER}" "mympdos-mpd-stable-${MPDVER}"
-  tar -czf mympdos-mpd-stable.tar.gz "mympdos-mpd-stable-${MPDVER}"
-  rm -fr "mympdos-mpd-stable-${MPDVER}"
+  su build -c "wget http://www.musicpd.org/download/mpd/0.21/mpd-${B_MPD_STABLE_VER}.tar.xz"
+  tar -xf "mpd-${B_MPD_STABLE_VER}.tar.xz"
+  rm "mpd-${B_MPD_STABLE_VER}.tar.xz"
+  mv "mpd-${B_MPD_STABLE_VER}" "mympdos-mpd-stable-${B_MPD_STABLE_VER}"
+  tar -czf mympdos-mpd-stable.tar.gz "mympdos-mpd-stable-${B_MPD_STABLE_VER}"
+  rm -fr "mympdos-mpd-stable-${B_MPD_STABLE_VER}"
   su build -c "abuild checksum"
   su build -c "abuild -r"
   cd ..
